'qcreport' subpackage renamed to 'qc'.
