+++ auto_process_ngs/utils.py
@@ -34,10 +34,10 @@ import bcftbx.TabFile as TabFile
 import bcftbx.JobRunner as JobRunner
 import bcftbx.Pipeline as Pipeline
 import bcftbx.utils as bcf_utils
-from qcreport.illuminaqc import QCReporter
-from qcreport.illuminaqc import QCSample
-from qcreport.illuminaqc import expected_qc_outputs
-from qcreport.illuminaqc import check_qc_outputs
+from qc.illuminaqc import QCReporter
+from qc.illuminaqc import QCSample
+from qc.illuminaqc import expected_qc_outputs
+from qc.illuminaqc import check_qc_outputs
 
 #######################################################################
 # Classes
